+++ tests/Integration/Parser/DisplayMapTest.php
@@ -151,4 +151,25 @@ class DisplayMapTest extends TestCase {
 		);
 	}
 
+	public function testWhenLocationHasVisitedIconModifierWithNamespacePrefix_thePrefixGetsRemoved() {
+		$this->assertContains(
+			'"visitedicon":"VisitedIcon.png"',
+			$this->parse( '{{#display_map:1,1~title~text~icon~group~inline label~File:VisitedIcon.png}}' )
+		);
+	}
+
+	public function testWhenVisitedIconParameterIsProvidedWithNamespacePrefix_thePrefixGetsRemoved() {
+		$this->assertContains(
+			'"visitedicon":"VisitedIcon.png"',
+			$this->parse( '{{#display_map:1,1|visitedicon=File:VisitedIcon.png}}' )
+		);
+	}
+
+	public function testWhenLocationHasIconModifierWithNamespacePrefix_thePrefixGetsRemoved() {
+		$this->assertContains(
+			'"icon":"Icon.png"',
+			$this->parse( '{{#display_map:1,1~title~text~File:Icon.png}}' )
+		);
+	}
+
 }